Self-cast fix and what I mean by stagnating
Self-cast spells no longer crash the game
Also, when I say "focus stagnates" I mean that, if the "addAmount" is
below 1, the way the formula works makes focus regen TOO low at low
values.
From 300 to 400 = .7 seconds
From 200 to 200 = 51.6 seconds
I'm all for the concept, but right now it's just unforgiving.
I've changed it back to what I put for the sake of testing things for
now; feel free to change it back if you so desire